Handle non-present files for computing cache key
Currently plugin fails when asked (via `cache-on` opt) to use file that doesn't exist for cache:

```
sha1sum: package-lock.json: No such file or directory
/var/lib/buildkite-agent/plugins/github-com-seek-oss-docker-ecr-cache-buildkite-plugin-v0-0-3/hooks/pre-command: line 34: sha: unbound variable
```

However, there're use cases when it's not known if files be there or not. E.g. commonly used pipelines that consume this plugin, that expect either package-lock.json or yarn.lock be present.

There still will a warning in stderr, but plugin will gracefully handle this situation.