Ignore errors when setting ECE lifecycle policy
Because the previous `upsert_ecr` didn't perform any operations if the
ECR repository existed it was possible to "bring your own" ECR
repository. For example, this build:

https://buildkite.com/arret/arret/builds/490#9ff9b724-80e2-40d2-a835-eecff8885773

Used a clickopsed ECR with a custom lifecycle policy. This allowed the
agent to have `ECRAccessPolicy: PowerUser` to limit its permissions.
However, the change in #9 broke this by unconditionally updating the
policy in an attempt to fix previously autocreated repositories.

Fix by just ignoring the error. It will still appear in the output but
it won't break the build.