Use subdirectory (if any) of the Dockerfile as the build context
This enables support for Dockerfiles that aren't in the root directory (eg: in monorepos). By using the subdirectory containing the Dockerfile as the build context, .dockerignore and COPY commands work relative to the Dockerfile, rather than the root of the repo.

This is backward incompatible change that will break COPY commands that are currently specified relative to the repo root, rather than the Dockerfile subdirectory. I've checked existing usages and haven't been able to find any such cases.